Items from the Backend/MongoDB with lon and lat will now get rendered onto the map under MapPage.js. A problem is that they  now only render after a second reload, because when rendering the map the backend is not fast enough to bring the new element back in the first request.
