+++ backend/src/main/java/de/th/ecobobackend/model/dto/EcoElementDto.java
@@ -24,6 +24,6 @@ public class EcoElementDto {
     Boolean isInBochum;
     Boolean certificate1;
     Boolean certificate2;
-    String lon;
-    String lat;
+    Double lon;
+    Double lat;
 }
