+++ frontend/src/main/ListPage.js
@@ -7,27 +7,20 @@ import {getEcoElements} from "./controller/EcoElementController";
 import FoodStoreList from "./list-route/FoodStoreList";
 import RestaurantList from "./list-route/RestaurantList";
 import FairShopList from "./list-route/FairShopList";
-import Fab from '@material-ui/core/Fab';
-import AddIcon from '@material-ui/icons/Add';
+import AddItemButton from "./designElements/buttons/AddItemButton";
 
 export default function ListPage() {
 
     const history = useHistory();
     const {ecoElements, setEcoElements} = useContext(EcoElementContext);
 
-
-
     useEffect(() => {
         getEcoElements().then(setEcoElements);
-    }, [setEcoElements]);
+    }, []);
 
 
     function handleEditElement() {
         history.push("/404");
-    }
-
-    function handleAddElementButton() {
-        history.push("/bo/addElement");
     }
 
     return(
@@ -42,11 +35,7 @@ export default function ListPage() {
 
                 {/* the brs are necessary at the moment to keep the full list visible when scrolling */}
                 <br/>
-                <StyledActionButton>
-                    <Fab color="primary" aria-label="add" size="small" onClick={handleAddElementButton}>
-                        <AddIcon />
-                    </Fab>
-                </StyledActionButton>
+                <AddItemButton/>
                 <br/>
                 <br/>
                 <br/>
@@ -84,17 +73,5 @@ const StyledBottomDiv = styled.div`
   z-index: 10;
 `
 
-const StyledActionButton = styled.div`
-  margin: 0;
-  padding: 0;
-  align-self: flex-start;
-  background: transparent;
-  border: none;
-  font: inherit;
-  color: inherit;
-  position: fixed;
-  right: 25px;
-  bottom: 15px;
-  z-index: 100;
-`
+
